fix(plugins): renamed types.py to plugin_types.py due to conflict in pytest
