+++ mcpgateway/plugins/framework/plugin_types.py
